// indigov-challenge/src/types/constituent.type.ts
export type Constituent = {
    id: number;
    name: string;
    phoneNumber: string;
    emailAddress: string;
    phoneNumberVerified: boolean;
    emailVerified: boolean;
    address1: string;
    adress2: string;
    city: string;
    state: string;
    zipCode: string;
    isRegistered: boolean;
}